only processCoadd/forcedPhotCoadd on existing data
Added an ExistingCoaddDataIdContainer which tosses data references
that don't exist.  Using this in ProcessCoaddTask and ForcedPhotCoaddTask
means we don't get unnecessary backtraces when we know the data doesn't
exist.

Note this is essentially a cherry-pick of HSC commit
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/1ab95d8fd0aadf5be3349a901392b1ac4d4d3ec6

This class is in coadd_utils (as oppoesed to pipe_tasks) here as it is
called from within meas_base which cannot count on having pipe_tasks
built when run (e.g. all tests in a Buildbot would fail).