CoaddDataIdContainer: use given dataset instead of deepCoadd
The use of deepCoadd (well, the "deep" part was configurable) was
hard-coded, but we now write background-subtracted coadds as
deepCoadd_calexp since HSC-1311.  Changed
CoaddDataIdContainer.makeDataRefList to respect the provided dataset
type instead of always grabbing deepCoadd.  However, the skymap
doesn't have the same flexibility, and is always deepCoadd_skyMap,
so simplified the parameter list.

Conflicts:
	python/lsst/pipe/tasks/coaddBase.py

This is essentially a cherry-pick of HSC-1366's:
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/ea5e1ec609da31063f5c20a5f7f0943fcc0cf718

Noting that the coaddDataIdContainer lives in coadd_utils in the
LSST stack.

Note also that HSC-1311 will be ported subsequently.