Add TractDataIdContainer
It is difficult to make a data reference that merely points to an entire
tract: there is no data product solely at the tract level.  This adds
a TractDataIdContainer which generates a list of data references for
patches within a given tract.

Note that TractDataIdContainer was introduced in HSC's standalone commit:
https://github.com/HyperSuprime-Cam/hscPipe/commit/4a8465424316019d8f87fcb6ac0e4f2d7309ad91

It was added to HSC's python/hsc/pipe/tasks/stack.py task which does
not exist in the LSST stack, so it was added to coadd_utils here in
the file which also contains CoaddDataIdContainer.